move listening thread and client code into client.py
